requirements.txt
# Deep Learning
torch==2.6.0
torchvision==0.21.0

# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Image Processing
Pillow==11.0.0

# Data Science
numpy==1.26.4
scikit-learn==1.6.0

# Utilities
werkzeug==3.1.3
